RescueGrid_KTK/CHATBOTSERVER/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
pyttsx3==2.90
langid==1.1.6
googletrans==3.1.0a0
google-generativeai==0.3.0
gunicorn==21.2.0